src/k1_walk_mujoco/robot/k1_defaults.py
from pathlib import Path
from typing import Any

import yaml

_CONFIG_PATH = Path(__file__).resolve().parents[3] / "configs" / "robot_k1_pd.yaml"


def load_robot_pd_config() -> dict[str, Any]:
    with _CONFIG_PATH.open("r", encoding="utf-8") as f:
        return yaml.safe_load(f)
