pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "k1-walk-mujoco"
version = "0.1.0"
description = "Booster K1 walking environment in MuJoCo with PD position-target control"
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "k1-walk-mujoco contributors" }]
dependencies = [
  "gymnasium>=0.29,<1.1",
  "mujoco>=3.1",
  "numpy>=1.24",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.8",
  "pytest>=8.0",
  "ruff>=0.3",
]
sb3 = [
  "stable-baselines3>=2.3",
  "torch>=2.2",
]
cleanrl = [
  "torch>=2.2",
  "gymnasium>=0.29,<1.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
